# specs/007-weekly-plan/contracts/weekly-plans.openapi.yaml
openapi: 3.1.0
info:
  title: Meal Pilot Weekly Plans API
  version: 1.0.0
  description: >
    Organizer-facing household WeeklyPlan library. Each plan is one Monday
    week-start with up to seven day slots (one meal per day) linking to
    household Recipes. Filled slots use status pending | approved | rejected
    (default pending on assign/replace). Same Recipe may appear on multiple
    days. Slot edits are per-slot assign/replace, clear, and status actions.
    Recipe library delete is blocked while a slot references that Recipe
    (RECIPE_IN_USE). GenerateWeeklyMeals, alternative suggestions, multi-meal
    types per day, and BuildGroceryList are out of scope.
servers:
  - url: http://localhost:3000
    description: Local development
paths:
  /weekly-plans:
    get:
      operationId: listWeeklyPlans
      summary: List household weekly plans (newest week-start first)
      responses:
        "200":
          description: Plans ordered by weekStartDate descending
          content:
            application/json:
              schema:
                type: object
                required: [items, maxWeeklyPlans]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/WeeklyPlanSummary"
                  maxWeeklyPlans:
                    type: integer
                    const: 104
    post:
      operationId: createWeeklyPlan
      summary: >
        Create a weekly plan for a Monday week-start (optional initial slots;
        empty week-start-only create allowed)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWeeklyPlanRequest"
      responses:
        "201":
          description: Weekly plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"
        "400":
          description: >
            Validation failure (missing weekStartDate, non-Monday date, invalid
            day, duplicate day in slots payload)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Recipe not found in household (when slots provided)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >
            Duplicate week (WEEKLY_PLAN_CONFLICT) or household already at 104
            plans (WEEKLY_PLAN_LIBRARY_FULL)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /weekly-plans/{weeklyPlanId}:
    parameters:
      - $ref: "#/components/parameters/WeeklyPlanId"
    get:
      operationId: getWeeklyPlan
      summary: Get weekly plan detail (seven days Monday–Sunday)
      responses:
        "200":
          description: Full weekly plan with materialized day slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"
        "404":
          description: Weekly plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteWeeklyPlan
      summary: Permanently remove a weekly plan and its slots
      responses:
        "204":
          description: Deleted
        "404":
          description: Weekly plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /weekly-plans/{weeklyPlanId}/slots/{day}:
    parameters:
      - $ref: "#/components/parameters/WeeklyPlanId"
      - $ref: "#/components/parameters/Day"
    put:
      operationId: assignWeeklyPlanSlot
      summary: >
        Assign or replace Recipe on one day (sets status to pending; other days
        unchanged)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignSlotRequest"
      responses:
        "200":
          description: Weekly plan after slot assign/replace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"
        "400":
          description: Validation failure (missing recipeId or unexpected fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Weekly plan or Recipe not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: clearWeeklyPlanSlot
      summary: >
        Clear one day slot (other days unchanged; idempotent if day already
        empty; returns full WeeklyPlan)
      responses:
        "200":
          description: Weekly plan after slot clear (including idempotent empty clear)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"
        "404":
          description: Weekly plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /weekly-plans/{weeklyPlanId}/slots/{day}/status:
    parameters:
      - $ref: "#/components/parameters/WeeklyPlanId"
      - $ref: "#/components/parameters/Day"
    put:
      operationId: setWeeklyPlanSlotStatus
      summary: >
        Set approval status on a filled slot only (does not change Recipe)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetSlotStatusRequest"
      responses:
        "200":
          description: Weekly plan after status change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyPlan"
        "400":
          description: >
            Validation failure (empty slot, invalid status, or unexpected
            recipeId)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Weekly plan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  parameters:
    WeeklyPlanId:
      name: weeklyPlanId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Day:
      name: day
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Weekday"
  schemas:
    Weekday:
      type: string
      enum:
        - monday
        - tuesday
        - wednesday
        - thursday
        - friday
        - saturday
        - sunday
    SlotStatus:
      type: string
      enum:
        - pending
        - approved
        - rejected
      description: Canonical names; not "suggested"
    MealSlot:
      type: object
      required: [day, recipeId, recipeTitle, status]
      properties:
        day:
          $ref: "#/components/schemas/Weekday"
        recipeId:
          type: ["string", "null"]
          format: uuid
          description: Null when the day is empty
        recipeTitle:
          type: ["string", "null"]
          description: Current Recipe title when filled; null when empty
        status:
          oneOf:
            - $ref: "#/components/schemas/SlotStatus"
            - type: "null"
          description: Null when the day is empty
    WeeklyPlanSummary:
      type: object
      required:
        - id
        - householdId
        - weekStartDate
        - filledSlotCount
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        weekStartDate:
          type: string
          format: date
          description: Monday ISO date YYYY-MM-DD
        filledSlotCount:
          type: integer
          minimum: 0
          maximum: 7
          description: Count of days with a Recipe assigned (convenience field)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WeeklyPlan:
      type: object
      required:
        - id
        - householdId
        - weekStartDate
        - slots
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        weekStartDate:
          type: string
          format: date
          description: Monday ISO date YYYY-MM-DD; immutable after create
        slots:
          type: array
          minItems: 7
          maxItems: 7
          description: Always Monday through Sunday in calendar order
          items:
            $ref: "#/components/schemas/MealSlot"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateWeeklyPlanRequest:
      type: object
      required: [weekStartDate]
      additionalProperties: false
      properties:
        weekStartDate:
          type: string
          format: date
          description: Must be a Monday (UTC calendar)
        slots:
          type: array
          maxItems: 7
          description: Optional initial assignments; omit or [] for empty plan
          items:
            type: object
            required: [day, recipeId]
            additionalProperties: false
            properties:
              day:
                $ref: "#/components/schemas/Weekday"
              recipeId:
                type: string
                format: uuid
    AssignSlotRequest:
      type: object
      required: [recipeId]
      additionalProperties: false
      properties:
        recipeId:
          type: string
          format: uuid
      description: >
        Do not include status; assign/replace always sets status to pending.
    SetSlotStatusRequest:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status:
          $ref: "#/components/schemas/SlotStatus"
      description: Do not include recipeId; presence is VALIDATION_ERROR.
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - WEEKLY_PLAN_CONFLICT
            - WEEKLY_PLAN_LIBRARY_FULL
            - RECIPE_IN_USE
            - NOT_FOUND
        message:
          type: string